requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==3.0.10
twilio==7.17.0
python-dotenv==0.19.0
